requirements.txt
discord.py>=2.3.2
aiohttp>=3.9.0
python-dotenv>=1.0.0
aiosqlite>=0.19.0
wikipedia>=1.4.0
pyshorteners>=1.0.1
yt-dlp>=2023.0.0
PyNaCl>=1.5.0
